feat: Add password reset via email functionality
