+++ .sequelizerc
@@ -1,8 +1,9 @@
-const path = require('path');
 
-module.exports = {
-  'config': path.resolve('src/config', 'config.js'),
-  'models-path': path.resolve('src/', 'models'),
-  'seeders-path': path.resolve('src/', 'seeders'),
-  'migrations-path': path.resolve('src/', 'migrations')
+const {resolve} = require('path')
+
+module.exports={
+  'config': resolve('src/config', 'config.js'),
+  'models-path': resolve('src/', 'models'),
+  'seeders-path': resolve('src/', 'seeders'),
+  'migrations-path': resolve('src/', 'migrations')
 };
