+++ package.json
@@ -53,6 +53,7 @@
     "babel-plugin-istanbul": "^6.1.1",
     "bcrypt": "^5.1.0",
     "bcryptjs": "^2.4.3",
+    "body-parser": "^1.20.2",
     "chai": "^4.3.7",
     "chai-http": "^4.3.0",
     "dotenv": "^16.0.3",
