+++ src/Database/config/config.js
@@ -1,0 +1,12 @@
+/* eslint-disable linebreak-style */
+require('dotenv').config();
+
+module.exports = {
+  testing: {
+    username: process.env.DB_USERNAME,
+    password: process.env.DB_PASSWORD,
+    database: process.env.DB_NAME,
+    host: process.env.DB_HOST,
+    dialect: 'postgres'
+  }
+};
