--- src/swagger/Password-Reset.swagger.yaml
@@ -1,71 +1,0 @@
-swagger: "2.0"
-info:
-  version: "1.0.0"
-  title: "RESET PASSWORD API"
-  description: "RESET PASSWORD"
-basePath: "/"
-schemes:
-  - "http"
-consumes:
-  - "application/json"
-produces:
-  - "application/json"
-
-paths:
-  /requestPasswordReset:
-    post:
-      tags:
-        - "User Send Request to rest"
-      summary: "Request Password Reset"
-      description: "Endpoint for requesting password reset"
-      parameters:
-        - in: body
-          name: body
-          description: "User email address"
-          required: true
-          schema:
-            type: object
-            properties:
-              email:
-                type: string
-                example: "example@gmail.com"
-      responses:
-        200:
-          description: "Password reset request sent successfully"
-          schema:
-            type: object
-            properties:
-              message:
-                type: string
-                example: "Password reset request sent successfully"
-
-  /reset-password:
-    post:
-      tags:
-        - "Validate token and Reset"
-      summary: "Reset Password"
-      description: "Endpoint for resetting password"
-      parameters:
-        - in: body
-          name: body
-          description: "User reset token and new password"
-          required: true
-          schema:
-            type: object
-            properties:
-              resetToken:
-                type: string
-                example: "75d04b786b3bc2b19668812e45d7bc09a3858c22"
-              newPassword:
-                type: string
-                example: "Password@12"
-      responses:
-        200:
-          description: "Password reset successfully"
-          schema:
-            type: object
-            properties:
-              message:
-                type: string
-                example: "Password reset successfully"
-
